# image-crud/src/main/resources/api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080/api/v1
  description: Generated server url
tags:
- name: Image Controller
  description: Базовый CRUD API для работы с картинками
- name: Authentication Controller
  description: "API для аутентификации, регистрации и обновления токенов"
paths:
  /image:
    post:
      tags:
      - Image Controller
      summary: Загрузка нового изображения в систему
      description: |-
        В рамках данного метода необходимо:
        1. Провалидировать файл. Максимальный размер файла - 10Мб, поддерживаемые расширения - png, jpeg.
        2. Загрузить файл в S3 хранилище.
        3. Сохранить в БД мета-данные файла - название; размер; ИД файла в S3; ИД пользователя, которому файл принадлежит.
      operationId: uploadImage
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "500":
          description: Непредвиденная ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
        "200":
          description: Успех выполнения операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadImageResponse'
        "400":
          description: Файл не прошел валидацию
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
      security:
      - Auth: []
  /auth/signup:
    post:
      tags:
      - Authentication Controller
      summary: Регистрация аккаунта
      operationId: signUp
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountDto'
        required: true
      responses:
        "400":
          description: Некорректный формат данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
        "200":
          description: Регистрация успешна
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
        "401":
          description: Некорректные реквизиты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
  /auth/refresh:
    post:
      tags:
      - Authentication Controller
      summary: Обновление токена
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JwtRefreshRequest'
        required: true
      responses:
        "400":
          description: Некорректный формат данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
        "200":
          description: Создан новый токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
        "403":
          description: Некорректные реквизиты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
  /auth/login:
    post:
      tags:
      - Authentication Controller
      summary: Вход в аккаунт
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JwtRequest'
        required: true
      responses:
        "200":
          description: Вход успешен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
        "400":
          description: Некорректный формат данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
        "401":
          description: Некорректные реквизиты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
  /images:
    get:
      tags:
      - Image Controller
      summary: "Получение списка изображений, которые доступны пользователю"
      description: |-
        В рамках данного метода необходимо:
        1. Получить мета-информацию о всех изображениях, которые доступны пользователю
      operationId: getImages
      responses:
        "500":
          description: Непредвиденная ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
        "200":
          description: Успех выполнения операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetImagesResponse'
      security:
      - Auth: []
  /image/{image-id}:
    get:
      tags:
      - Image Controller
      summary: Скачивание файла по ИД
      description: |-
        В рамках данного метода необходимо:
        1. Проверить, есть ли такой файл в системе.
        2. Проверить, доступен ли данный файл пользователю.
        3. Скачать файл.
      operationId: downloadImage
      parameters:
      - name: image-id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "404":
          description: Файл не найден в системе или недоступен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
        "500":
          description: Непредвиденная ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
        "200":
          description: Успех выполнения операции
          content:
            '*/*':
              schema:
                type: string
                format: binary
      security:
      - Auth: []
    delete:
      tags:
      - Image Controller
      summary: Удаление файла по ИД
      description: |-
        В рамках данного метода необходимо:
        1. Проверить, есть ли такой файл в системе.
        2. Проверить, доступен ли данный файл пользователю.
        3. Удалить файл.
      operationId: deleteImage
      parameters:
      - name: image-id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "500":
          description: Непредвиденная ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
        "200":
          description: Успех выполнения операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
        "404":
          description: Файл не найден в системе или недоступен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
      security:
      - Auth: []
components:
  schemas:
    UiSuccessContainer:
      required:
      - success
      type: object
      properties:
        success:
          type: boolean
          description: Признак успеха
        message:
          type: string
          description: Сообщение об ошибке
    UploadImageResponse:
      required:
      - imageId
      type: object
      properties:
        imageId:
          type: string
          description: ИД файла
          format: uuid
    CreateAccountDto:
      required:
      - confirmPassword
      - password
      - username
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
    JwtResponse:
      type: object
      properties:
        accountId:
          type: string
        username:
          type: string
        accessToken:
          type: string
        refreshToken:
          type: string
    JwtRefreshRequest:
      required:
      - token
      type: object
      properties:
        token:
          type: string
    JwtRequest:
      required:
      - password
      - username
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    GetImagesResponse:
      required:
      - images
      type: object
      properties:
        images:
          type: array
          items:
            $ref: '#/components/schemas/Image'
    Image:
      required:
      - filename
      - size
      type: object
      properties:
        imageId:
          type: string
          description: ИД файла
          format: uuid
        filename:
          type: string
          description: Название изображения
        size:
          type: integer
          description: Размер файла в байтах
          format: int32
  securitySchemes:
    Auth:
      type: http
      in: header
      scheme: bearer
      bearerFormat: JWT